Refactors toll sort classes
Adds sorting to favorite tolls
