+++ app/src/main/java/gov/wa/wsdot/android/wsdot/util/Utils.java
@@ -135,5 +135,4 @@ public class Utils {
             return arrayList;
         }
     }
-
 }
